Move AssaySpecimen and AssaySpecimenVisit tables to the study design query schema
