+++ study/api-src/org/labkey/api/studydesign/query/StudyDesignQuerySchema.java
@@ -43,6 +43,9 @@ public class StudyDesignQuerySchema extends SimpleUserSchema implements UserSche
     public static final String TREATMENT_VISIT_MAP_TABLE_NAME = "TreatmentVisitMap";
     public static final String OBJECTIVE_TABLE_NAME = "Objective";
 
+    public static final String ASSAY_SPECIMEN_TABLE_NAME = "AssaySpecimen";
+    public static final String ASSAY_SPECIMEN_VISIT_TABLE_NAME = "AssaySpecimenVisit";
+
     protected Study _study;
     private Role _contextualRole;
     private UserSchema _parentSchema;
@@ -68,7 +71,9 @@ public class StudyDesignQuerySchema extends SimpleUserSchema implements UserSche
             TREATMENT_TABLE_NAME,
             PERSONNEL_TABLE_NAME,
             TREATMENT_VISIT_MAP_TABLE_NAME,
-            OBJECTIVE_TABLE_NAME
+            OBJECTIVE_TABLE_NAME,
+            ASSAY_SPECIMEN_TABLE_NAME,
+            ASSAY_SPECIMEN_VISIT_TABLE_NAME
     );
 
     private StudyDesignQuerySchema(UserSchema studySchema, Study study, @Nullable Role contextualRole)
@@ -203,6 +208,14 @@ public class StudyDesignQuerySchema extends SimpleUserSchema implements UserSche
         if (OBJECTIVE_TABLE_NAME.equalsIgnoreCase(name))
         {
             return new StudyObjectiveTable(this, cf);
+        }
+        if (ASSAY_SPECIMEN_TABLE_NAME.equalsIgnoreCase(name))
+        {
+            return new AssaySpecimenTable(this, cf);
+        }
+        if (ASSAY_SPECIMEN_VISIT_TABLE_NAME.equalsIgnoreCase(name))
+        {
+            return new AssaySpecimenVisitTable(this, cf);
         }
 
         return null;
